+++ tests/function/test_check_all.py
@@ -44,3 +44,32 @@ def test_missing_space_before_act(function):
     assert excinfo.value.line_number == 4
     assert excinfo.value.offset == 4
     assert excinfo.value.text == 'AAA03 expected 1 blank line before Act block, found none'
+
+
+@pytest.mark.parametrize(
+    'code_str',
+    [
+        '''
+def test(file_resource):
+    file_resource.connect()
+
+    result = file_resource.retrieve()
+    assert result.success is True
+''',
+        '''
+def test_push(queue):
+    item = Document()
+
+    queue.push(item)  # act
+    assert queue.pop() == item
+''',
+    ],
+    ids=['no line before assert', 'no line before assert with marked act'],
+)
+def test_missing_space_before_assert(function):
+    with pytest.raises(ValidationError) as excinfo:
+        function.check_all()
+
+    assert excinfo.value.line_number == 6
+    assert excinfo.value.offset == 4
+    assert excinfo.value.text == 'AAA04 expected 1 blank line before Assert block, found none'